chore: rename rulesets-new -> rulesets
