+++ src/rulesets/json-api-rules/__tests__/compound-document-rules.test.ts
