+++ src/rulesets/json-api-rules/__tests__/content-type-rules.test.ts
