+++ src/rulesets/json-api-rules/__tests__/disallow-delete-and-post-singleton-rule.test.ts
