+++ src/rulesets/json-api-rules/__tests__/resource-object-rules.test.ts
