+++ src/rulesets/json-api-rules/compound-document-rules.ts
