+++ src/rulesets/json-api-rules/content-type-rules.ts
