+++ src/rulesets/json-api-rules/disallow-delete-and-post-singleton-rule.ts
